refactor: add new view component for todos list
